+++ src/EpiInvert_main.cpp
@@ -5,7 +5,70 @@ using namespace std;
  
 #include "EpiInvertCore_q_variable.h"
 
-
+ // [[Rcpp::export]]
+ List EpiInvertForecastC(
+     NumericVector i_restored,
+     String last_incidence_date,
+     NumericVector q_bias,
+     NumericMatrix i_restored_database
+ )
+ {
+   string last_incidence_dateC=string(last_incidence_date.get_cstring());/** DATE OF THE LAST DATA IN THE FORMAT YYYY-MM-DD */;
+   
+   vector< vector<double> > M(i_restored_database.nrow(),vector<double>(i_restored_database.ncol(),0.));
+   //printf("%d %d\n",i_restored_database.ncol(),i_restored_database.nrow());
+   for(int i=0;i<(int) M.size();i++){
+     for(int j=0;j<(int) M[i].size();j++){
+       M[i][j]=i_restored_database(i,j); 
+     }
+   }
+   
+   vector<double> ir(i_restored.size()),q(q_bias.size());
+   for(int k=0;k<(int) ir.size();k++) ir[k]=i_restored[k];
+   for(int k=0;k<(int) q.size();k++) q[k]=q_bias[k];
+   
+   double lambda=176.8;
+   double mu=0.035;
+   vector <double> CI50,CI75,CI90,CI95,i0_forecast;
+   vector<string> dates;
+   
+   //printf("ir.size()=%d, q.size()=%d, M.size()=%d\n",ir.size(),q.size(),M.size()); 
+   //printf("ir[0]=%lf,q[0]=%lf,M[0][0]=%lf\n",ir[0],q[0],M[0][0]); 
+   
+   //return List::create(); 
+   
+   vector<double> v=IncidenceForecastByLearning(
+     ir,
+     last_incidence_dateC,
+     q,
+     M,
+     lambda,
+     mu,
+     CI50,
+     CI75,
+     CI90,
+     CI95,
+     i0_forecast,
+     dates
+   );
+   
+   
+   //printf("M[0][0]=%lf M[0][1]=%lf, M[1][0]=%lf\n",M[0][0],M[0][1],M[1][0]);
+   //printf("%lf  %lf\n",N(0,0),N(0,1));
+   
+   List results = List::create(
+     Named("i_restored_forecast") = v ,
+     Named("i_restored_forecast_CI50") = CI50 ,
+     Named("i_restored_forecast_CI75") = CI75 ,
+     Named("i_restored_forecast_CI90") = CI90 ,
+     Named("i_restored_forecast_CI95") = CI95 ,
+     Named("dates")  = dates,
+     Named("i_original_forecast")  = i0_forecast
+   );
+   return results;
+   
+ }
+     
 // [[Rcpp::export]]
 List EpiInvertC(
     NumericVector i_original0,
@@ -18,7 +81,8 @@ List EpiInvertC(
     double sd_si=5.667547,
     double shift_si=-5.,
     double Rt_regularization_weight=5.,
-    double seasonality_regularization_weight=5.
+    double seasonality_regularization_weight=5.,
+    bool incidence_weekly_aggregated=false
 ){
   clock_t t=clock();
   
@@ -113,7 +177,8 @@ List EpiInvertC(
     Rt_regularization_weight /** REGULARIZATION WEIGHT PARAMETER OF EpiInvert METHOD (DEFAULT VALUE: 5)*/,
     seasonality_regularization_weight /** WEIGHT PARAMETER OF THE REGULARIZATION  TERM FOR THE SEASONALITY q (DEFAULT VALUE 5) */,
     max_time_interval /** MAX SIZE OF THE INCIDENCE DATA USED TO COMPUTE Rt (DEFAULT VALUE: 9999). THIS PARAMETER IS USED TO REDUCE HE COMPUTATIONAL COST OF THE ALGORITHM WHEN WE ARE JUST INTERESTED IN THE LAST PART OF THE SEQUENCE */,
-    NweeksToKeepIncidenceSum /** WE CONSTRAINT ALL THE ESTIMATED INCIDENCE CURVE TO KEEP THE ADDITION OF THE ORIGINAL INCIDENCE IN INTERVALS OF SIZE NweeksToKeepIncidenceSum*7 DAYS*/
+    NweeksToKeepIncidenceSum /** WE CONSTRAINT ALL THE ESTIMATED INCIDENCE CURVE TO KEEP THE ADDITION OF THE ORIGINAL INCIDENCE IN INTERVALS OF SIZE NweeksToKeepIncidenceSum*7 DAYS*/,
+    incidence_weekly_aggregated /** IF TRUE, EACH INCIDENCE VALUE CORRESPONDS TO THE LAST 7-DAY AGGREGATED INCIDENCE */
   );